split program into two parts
