+++ cs/MarkdownProcessing/MarkdownProcessing.csproj
@@ -55,8 +55,9 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Context.cs" />
-    <Compile Include="MarkdownProcessing_Should\Context_Should.cs" />
+    <Compile Include="TokenParser.cs" />
+    <Compile Include="MarkdownConverter.cs" />
+    <Compile Include="MarkdownProcessing_Should\MarkdownConverter_Should.cs" />
     <Compile Include="MarkdownProcessing_Should\Token_Should.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Token.cs" />
