+++ cs/MarkdownProcessing/MarkdownProcessing_Should/MarkdownConverter_Should.cs
@@ -5,62 +5,42 @@ using NUnit.Framework;
 namespace MarkdownProcessing.MarkdownProcessing_Should
 {
     [TestFixture]
-    public class Context_Should
+    public class MarkdownConverter_Should
     {
-        [Test]
-        public void Context_ShouldParsePlaintTextTag()
-        {
-            var expected = "<p>Hello world</p>";
-            var actual = new Context("Hello world").ParseInputIntoTokens();
-            actual.Should().Be(expected);
-        }
-
-        [Test]
-        public void Context_ShouldParseSimpleTag()
-        {
-            var expected = "<strong><p>Hello world</p></strong>";
-            var actual = new Context("_Hello world_").ParseInputIntoTokens();
-            actual.Should().Be(expected);
-        }
-
         [Test]
         public void ConvertToken_ShouldConvertPlainTextToken()
         {
-            var context = new Context("Hello world");
-            context.ConvertToken(new SimpleToken("Hello world"));
-            context.output.ToString().Should().Be("Hello world");
+            var converter = new MarkdownConverter(new SimpleToken("Hello world"));
+            converter.ConvertToHtml().Should().Be("Hello world");
         }
 
         [Test]
         public void ConvertToken_ShouldConvertSimpleToken()
         {
-            var context = new Context("_Hello world_");
-            context.ConvertToken(new ComplicatedToken(TokenType.Bold)
+            var converter = new MarkdownConverter(new ComplicatedToken(TokenType.Bold)
             {
                 ChildTokens = new List<Token> {new SimpleToken("Hello world")}
             });
-            context.output.ToString().Should().Be("<strong>Hello world</strong>");
+            converter.ConvertToHtml().Should().Be("<strong>Hello world</strong>");
         }
 
         [Test]
         public void ConvertToken_ShouldConvertTokenWithDifferentTags()
         {
-            var context = new Context("#Hello world\n");
-            context.ConvertToken(new ComplicatedToken(TokenType.Header1)
+            var converter = new MarkdownConverter(new ComplicatedToken(TokenType.Header1)
             {
                 ChildTokens = new List<Token>
                 {
                     new SimpleToken("Hello world")
                 }
             });
-            context.output.ToString().Should().Be("<h1>Hello world</h1>");
+            converter.ConvertToHtml().Should().Be("<h1>Hello world</h1>");
         }
 
         [Test]
         public void ConvertToken_ShouldConvertSimpleTokenIntoParent()
         {
-            var context = new Context("_Hello world_");
-            context.ConvertToken(new ComplicatedToken(TokenType.Parent)
+            var converter = new MarkdownConverter(new ComplicatedToken(TokenType.Parent)
             {
                 ChildTokens = new List<Token>
                 {
@@ -70,14 +50,13 @@ namespace MarkdownProcessing.MarkdownProcessing_Should
                     }
                 }
             });
-            context.output.ToString().Should().Be("<p><strong>Hello world</strong></p>");
+            converter.ConvertToHtml().Should().Be("<p><strong>Hello world</strong></p>");
         }
 
         [Test]
         public void ConvertToken_ShouldConvertTwoSimpleTokensIntoParent()
         {
-            var context = new Context("_*Hello* world_");
-            context.ConvertToken(new ComplicatedToken(TokenType.Parent)
+            var converter = new MarkdownConverter(new ComplicatedToken(TokenType.Parent)
             {
                 ChildTokens = new List<Token>
                 {
@@ -94,14 +73,13 @@ namespace MarkdownProcessing.MarkdownProcessing_Should
                     }
                 }
             });
-            context.output.ToString().Should().Be("<p><strong><em>Hello</em> world</strong></p>");
+            converter.ConvertToHtml().Should().Be("<p><strong><em>Hello</em> world</strong></p>");
         }
 
         [Test]
         public void ConvertToken_ShouldConvertHeaderInnerTokensIntoParent()
         {
-            var context = new Context("#Hello *world*\n");
-            context.ConvertToken(new ComplicatedToken(TokenType.Parent)
+            var converter = new MarkdownConverter(new ComplicatedToken(TokenType.Parent)
             {
                 ChildTokens = new List<Token>
                 {
@@ -121,7 +99,7 @@ namespace MarkdownProcessing.MarkdownProcessing_Should
                     }
                 }
             });
-            context.output.ToString().Should().Be("<p><h1>Hello <strong>world</strong></h1></p>");
+            converter.ConvertToHtml().Should().Be("<p><h1>Hello <strong>world</strong></h1></p>");
         }
     }
 }
